Finished the logic for tracking fluctuating values as tied to specific events.
Before, I was thinking that we only cared about a threshold in which to trigger an event in the first place.
I realized that each arbitrary event will have it's own relationship to the fluctuating values and how those pertain to the threshold.
Expanded the Event.java pojo to fit my needs..

This commit introduces some pretty bad time complexity. I need to look into this and see what refactors I can make to address this.
More unit testing coverage is next.